backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
qrcode==7.4.2
pillow==10.1.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
cloudinary==1.36.0
python-jose[cryptography]==3.3.0
authlib==1.2.1
httpx==0.25.2
requests==2.31.0